feat(goid): Implement assembly-optimized goroutine ID extraction
Performance breakthrough: ~2200x faster GID extraction (1.73ns vs 3987ns)

New files:
- goid_amd64.s: Native x86-64 assembly using TLS access
- goid_arm64.s: Native ARM64 assembly using g register (R28)
- goid_fast.go: Go wrapper with offset logic (152 bytes for Go 1.23-1.25)
- goid_fallback.go: Fallback for unsupported platforms

Technical details:
- Build constraints: go1.23 && !go1.26 && (amd64 || arm64)
- Zero allocations on fast path (0 B/op)
- Automatic fallback to runtime.Stack parsing
- Zero external dependencies (strategic decision for Go runtime proposal)

Benchmarks:
- Assembly (amd64): 1.73 ns/op, 0 B/op
- runtime.Stack: 3987 ns/op, 64 B/op

Documentation updated: README, CHANGELOG, ROADMAP
