+++ go.mod
@@ -5,4 +5,4 @@ go 1.24.0
 // Pure-Go Race Detector
 // Replacing ThreadSanitizer C++ implementation with native Go
 
-require golang.org/x/mod v0.30.0
+require golang.org/x/mod v0.30.0 // Go module file parsing
