+++ scripts/pre-release-check.sh
@@ -79,8 +79,11 @@ fi
 echo ""
 
 # 4. Go vet (exclude external/)
+# Note: -unsafeptr=false is needed for goid_fast.go which uses intentional
+# uintptr→unsafe.Pointer conversion for runtime.g access (standard pattern
+# used by petermattis/goid, outrigdev/goid, and similar libraries)
 log_info "Running go vet..."
-VET_OUTPUT=$(go vet ./internal/... ./race/... ./cmd/... ./examples/... 2>&1)
+VET_OUTPUT=$(go vet -unsafeptr=false ./internal/... ./race/... ./cmd/... ./examples/... 2>&1)
 VET_EXIT=$?
 if [ $VET_EXIT -eq 0 ]; then
     log_success "go vet passed"
@@ -325,21 +328,31 @@ else
 fi
 echo ""
 
-# 9. Check minimal dependencies policy (Pure Go - NO dependencies!)
+# 9. Check dependencies policy (Minimal dependencies - known allowed list)
 log_info "Checking dependencies policy..."
 if [ -f "go.mod" ]; then
-    # Count non-stdlib dependencies (excluding indirect)
-    CORE_DEPS=$(grep -E "^\s+github.com|^\s+golang.org" go.mod | grep -v "// indirect" | wc -l)
+    # Allowed dependencies (with reasons):
+    # - golang.org/x/mod: Go module file parsing (for replace directive handling)
+    # - github.com/outrigdev/goid: Assembly-optimized goroutine ID extraction (~1000x faster)
+    ALLOWED_DEPS="golang.org/x/mod|github.com/outrigdev/goid"
 
-    # Expected: ZERO external dependencies (Pure Go, stdlib only!)
-    EXPECTED_DEPS=0
+    # Count dependencies not in allowed list (excluding indirect)
+    UNKNOWN_DEPS=$(grep -E "^\s+github.com|^\s+golang.org" go.mod | grep -v "// indirect" | grep -vE "$ALLOWED_DEPS" | wc -l)
+    TOTAL_DEPS=$(grep -E "^\s+github.com|^\s+golang.org" go.mod | grep -v "// indirect" | wc -l)
 
-    if [ "$CORE_DEPS" -eq 0 ]; then
-        log_success "Pure Go: Zero external dependencies ✅ (stdlib only)"
+    if [ "$UNKNOWN_DEPS" -eq 0 ]; then
+        if [ "$TOTAL_DEPS" -eq 0 ]; then
+            log_success "Pure Go: Zero external dependencies ✅ (stdlib only)"
+        else
+            log_success "Minimal dependencies: $TOTAL_DEPS allowed dependency(ies) ✅"
+            grep -E "^\s+github.com|^\s+golang.org" go.mod | grep -v "// indirect" | while read dep; do
+                echo "  • $dep"
+            done
+        fi
     else
-        log_error "Pure Go policy violated: $CORE_DEPS external dependencies found"
-        log_error "Race detector MUST use stdlib only (no CGO, no external deps)"
-        grep -E "^\s+github.com|^\s+golang.org" go.mod | grep -v "// indirect"
+        log_error "Unknown dependency found: $UNKNOWN_DEPS dependency(ies) not in allowed list"
+        log_error "Allowed: golang.org/x/mod, github.com/outrigdev/goid"
+        grep -E "^\s+github.com|^\s+golang.org" go.mod | grep -v "// indirect" | grep -vE "$ALLOWED_DEPS"
         ERRORS=$((ERRORS + 1))
     fi
 else
